Final DLR HTTP

+++ jasmin/managers/dlr.py
@@ -311,6 +311,8 @@ class DLRLookup(object):
             if len(dlr) > 0 and dlr['sc'] != connector_type:
                 raise DLRMapError('Found a dlr for msgid:%s with diffrent sc: %s' % (submit_sm_queue_id, dlr['sc']))
 
+            final_states = ['DELIVRD', 'EXPIRED', 'DELETED', 'UNDELIV', 'REJECTD']
+
             if connector_type == 'httpapi':
                 self.log.debug('There is a HTTP DLR request for msgid[%s] ...', msgid)
                 dlr_url = dlr['url']
@@ -337,9 +339,9 @@ class DLRLookup(object):
                                                                                err=pdu_dlr_err,
                                                                                text=pdu_dlr_text,
                                                                                method=dlr_method))
-
-                    self.log.debug('Removing HTTP dlr map for msgid[%s]', submit_sm_queue_id)
-                    yield self.redisClient.delete('dlr:%s' % submit_sm_queue_id)
+                    if dlr_level == 3 and pdu_dlr_status in final_states:
+                        self.log.debug('Removing HTTP dlr map for msgid[%s]', submit_sm_queue_id)
+                        yield self.redisClient.delete('dlr:%s' % submit_sm_queue_id)
             elif connector_type == 'smppsapi':
                 self.log.debug('There is a SMPPs mapping for msgid[%s] ...', msgid)
                 system_id = dlr['system_id']
@@ -353,7 +355,6 @@ class DLRLookup(object):
                 registered_delivery_receipt = dlr['rd_receipt']
 
                 success_states = ['ACCEPTD', 'DELIVRD']
-                final_states = ['DELIVRD', 'EXPIRED', 'DELETED', 'UNDELIV', 'REJECTD']
                 # Do we need to forward the receipt to the original sender ?
                 if ((pdu_dlr_status in success_states and
                              registered_delivery_receipt == 'SMSC_DELIVERY_RECEIPT_REQUESTED') or
